use BigDecimal for amount

+++ utility-billing-system/billing-service/src/main/java/com/utilitybilling/billingservice/dto/OutstandingBalanceResponse.java
@@ -1,9 +1,11 @@
 package com.utilitybilling.billingservice.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class OutstandingBalanceResponse {
-    private String consumerId;
-    private double outstandingAmount;
+	private String consumerId;
+	private BigDecimal outstandingAmount;
 }
